api_service/requirements.txt
fastapi
uvicorn
pandas==2.2.2
python-dotenv
mlflow==2.22.0
dagshub
dvc
python-multipart